Update script to install Chrome, 7zip and SumatraPDF
The build script was constantly updated and refined to install more
software and improve privacy settings. This version now ensures our new
system comes with the Chrome Browser (so we are never forced to use
Edge) and with a small PDF viewer.

The Chrome privacy settings are also enforced through GPOs.

The settings are hightly opinionated though but right now the build.bat
script is not designed to ask the user for choices.